Implement token authentication
This is adding a login mechanism inspired on the [kubernetes dashboard lofin view](https://github.com/kubernetes/dashboard/blob/master/docs/user/access-control/README.md#login-view).

Instead of providing a username and a password, Kiali asks for a token. The provided token is meant to be a service account token and Kiali will use it during the session of the user to make calls to the cluster API.

By using the provided token to make calls to the cluster API, the cluster's RBAC mechanisms takes effect and write actions in Kiali can be limited using Roles and RoleBindings.

Currently, in Kubernetes, there is a limitation: the resulting _accessible namespaces_ given the Kiali ConfigMap should all be, at least, readable via the kiali-viewer role. Kiali will fail if a user logs in using a token that does not have read privileges in, at least, one of the "accessible namespaces". This limitation is because the Kubernetes API always returns all namespaces regardless if the user does not have access to any of the resources on it. In contrast, in OpenShift Kiali takes advantage of the Projects API which truncates the list to the readable namespaces.

Despite the limitation with Kubernetes, the new Token auth mechanism allows to login with privileges to modify the cluster, or to login in view mode by taking advantage of the cluster RBAC.

See also kiali/kiali-ui#1572

Related #1139